requirement.txt
asgiref==3.4.1
bcrypt==3.2.0
boto3==1.20.4
botocore==1.23.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
configparser==5.1.0
cryptography==35.0.0
dj-database-url==0.5.0
Django==3.2.8
future==0.18.2
idna==3.3
jmespath==0.10.0
lxml==4.6.3
ncclient==0.6.12
paramiko==2.8.0
paramiko-expect==0.3.0
PyNaCl==1.4.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
scp==0.14.1
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
tornado==6.1
urllib3==1.26.7
whitenoise==5.3.0
